+++ tests/mock/mock.py
@@ -2,14 +2,16 @@ import socket
 from http.server import HTTPServer
 from threading import Thread
 
-from tests.mock.server.v2_1_1 import MockServerRequestHandler_v2_1_1
-from tests.mock.server.v2_1_2 import MockServerRequestHandler_v2_1_2
-from tests.mock.server.v2_2_1 import MockServerRequestHandler_v2_2_1
+
+
+
 from tests.mock.server.v2_2_2_3 import MockServerRequestHandler_v2_2_2_3
 from tests.mock.server.v2_2_3_3 import MockServerRequestHandler_v2_2_3_3
 from tests.mock.server.v2_3_2_0 import MockServerRequestHandler_v2_3_2_0
 from tests.mock.server.v2_3_3_0 import MockServerRequestHandler_v2_3_3_0
 from tests.mock.server.v2_3_5_3 import MockServerRequestHandler_v2_3_5_3
+from tests.mock.server.v2_3_7_6 import MockServerRequestHandler_v2_3_7_6
+
 
 HOST = 'localhost'
 
@@ -30,12 +32,6 @@ def get_mock_url(free_port):
 def start_mock_server(port, version):
     mockServerHandler = None
 
-    if version == '2.1.1':
-        mockServerHandler = MockServerRequestHandler_v2_1_1
-    if version == '2.1.2':
-        mockServerHandler = MockServerRequestHandler_v2_1_2
-    if version == '2.2.1':
-        mockServerHandler = MockServerRequestHandler_v2_2_1
     if version == '2.2.2.3':
         mockServerHandler = MockServerRequestHandler_v2_2_2_3
     if version == '2.2.3.3':
@@ -46,6 +42,8 @@ def start_mock_server(port, version):
         mockServerHandler = MockServerRequestHandler_v2_3_3_0
     if version == '2.3.5.3':
         mockServerHandler = MockServerRequestHandler_v2_3_5_3
+    if version == '2.3.7.6':
+        mockServerHandler = MockServerRequestHandler_v2_3_7_6
 
     if mockServerHandler is None:
         raise Exception('Could not create MockServer for version {}'.format(version))